# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is API for booking hotel with Xendit payment gateway",
        "title": "Staycation API",
        "termsOfService": "https://github.com/iqbalutomo/staycation",
        "contact": {
            "name": "API Support",
            "email": "muhlisiqbalutomo@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/iqbalutomo/staycation/blob/master/LICENSE"
        },
        "version": "0.0.1"
    },
    "host": "fc76-103-18-34-211.ngrok-free.app",
    "paths": {
        "/bookings": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Book room from hotel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Booking Room",
                "parameters": [
                    {
                        "description": "Booking details",
                        "name": "booking",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Booking"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/model.BookingSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        },
        "/users/deposit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deposit your balance for book room from hotel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Top Up Balance",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Deposit amount",
                        "name": "amount",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "description": "Login user for consume feature of API",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Login User",
                "parameters": [
                    {
                        "type": "string",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/model.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        },
        "/users/register": {
            "post": {
                "description": "Register a new user",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Register User",
                "parameters": [
                    {
                        "type": "string",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "phone",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "customer",
                            "hotel_owner"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "Customer",
                            "HotelOwner"
                        ],
                        "name": "role",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/model.RegisterSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Booking": {
            "type": "object",
            "required": [
                "check_in_date",
                "check_out_date"
            ],
            "properties": {
                "check_in_date": {
                    "type": "string"
                },
                "check_out_date": {
                    "type": "string"
                },
                "hotel_id": {
                    "type": "integer"
                },
                "id": {
                    "description": "gorm.Model",
                    "type": "integer"
                },
                "room_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "total_price": {
                    "type": "number"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.BookingResponse": {
            "type": "object",
            "properties": {
                "booking": {
                    "$ref": "#/definitions/model.Booking"
                },
                "invoice": {
                    "$ref": "#/definitions/model.Invoice"
                }
            }
        },
        "model.BookingSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.BookingResponse"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.Invoice": {
            "type": "object",
            "required": [
                "amount",
                "booking_id",
                "xendit_invoice_id"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "booking_id": {
                    "type": "integer"
                },
                "invoice_url": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "xendit_invoice_id": {
                    "type": "string"
                }
            }
        },
        "model.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                }
            }
        },
        "model.RegisterSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.User"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.User": {
            "type": "object",
            "required": [
                "email",
                "name",
                "phone",
                "role"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "description": "gorm.Model",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "enum": [
                        "customer",
                        "hotel_owner"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.UserRoleEnum"
                        }
                    ]
                }
            }
        },
        "model.UserRoleEnum": {
            "type": "string",
            "enum": [
                "customer",
                "hotel_owner"
            ],
            "x-enum-varnames": [
                "Customer",
                "HotelOwner"
            ]
        },
        "utils.APIError": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "error_code": {
                    "type": "string"
                },
                "message": {}
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "externalDocs": {
        "description": "Github",
        "url": "https://github.com/iqbalutomo"
    }
}